+++ common/src/main/kotlin/com/cobblemon/mod/common/pokemon/PokemonStats.kt
@@ -37,7 +37,7 @@ abstract class PokemonStats : Iterable<Map.Entry<Stat, Int>> {
     /** Run whenever anything changes. */
     var changeFunction: (PokemonStats) -> Unit = {}
 
-    private val stats = mutableMapOf<Stat, Int>()
+    protected val stats = mutableMapOf<Stat, Int>()
     private var emit = true
 
     fun doWithoutEmitting(action: () -> Unit) {
@@ -67,65 +67,5 @@ abstract class PokemonStats : Iterable<Map.Entry<Stat, Int>> {
 
     protected open fun canSet(stat: Stat, value: Int) = value in acceptableRange
 
-    fun saveToNBT(nbt: CompoundTag): CompoundTag {
-        this.stats.forEach { (stat, value) ->
-            // don't waste space if default
-            if (value != this.defaultValue) {
-                nbt.putShort(this.cleanStatIdentifier(stat.identifier), value.toShort())
-            }
-        }
-        return nbt
-    }
-
-    fun loadFromNBT(nbt: CompoundTag): PokemonStats {
-        stats.clear()
-        Stats.PERMANENT.forEach { stat ->
-            val identifier = this.cleanStatIdentifier(stat.identifier)
-            this[stat] = nbt.getShort(identifier).toInt().coerceIn(this.acceptableRange)
-        }
-        return this
-    }
-
-    fun saveToJSON(json: JsonObject): JsonObject {
-        this.stats.forEach { (stat, value) ->
-            // don't waste space if default
-            if (value != this.defaultValue) {
-                json.addProperty(this.cleanStatIdentifier(stat.identifier), value)
-            }
-        }
-        return json
-    }
-
-    fun loadFromJSON(json: JsonObject): PokemonStats {
-        stats.clear()
-
-        Stats.PERMANENT.forEach { stat ->
-            val identifier = this.cleanStatIdentifier(stat.identifier)
-            json.get(identifier)?.asInt?.coerceIn(this.acceptableRange)?.let { this.stats[stat] = it }
-        }
-        return this
-    }
-
-    fun saveToBuffer(buffer: RegistryFriendlyByteBuf) {
-        buffer.writeSizedInt(IntSize.U_BYTE, stats.size)
-        for ((stat, value) in stats) {
-            Cobblemon.statProvider.encode(buffer, stat)
-            buffer.writeSizedInt(IntSize.U_SHORT, value)
-        }
-    }
-
-    fun loadFromBuffer(buffer: RegistryFriendlyByteBuf) {
-        stats.clear()
-        repeat(times = buffer.readUnsignedByte().toInt()) {
-            val stat = Cobblemon.statProvider.decode(buffer)
-            val value = buffer.readUnsignedShort()
-            stats[stat] = value
-        }
-    }
-
     fun getOrDefault(stat: Stat) = this[stat] ?: this.defaultValue
-
-    // util to prevent unnecessary long identifiers, usually vanilla defaults to Minecraft but in our context defaulting to cobblemon makes more sense
-    private fun cleanStatIdentifier(identifier: ResourceLocation): String = identifier.toString().substringAfter("${Cobblemon.MODID}:")
-
 }